+++ install-local.ps1
@@ -242,8 +242,7 @@ $Dependencies = @(
     "Microsoft.Data.Sqlite.dll",
     "SQLitePCLRaw.batteries_v2.dll",
     "SQLitePCLRaw.core.dll",
-    "SQLitePCLRaw.provider.e_sqlite3.dll",
-    "Spectre.Console.dll"
+    "SQLitePCLRaw.provider.e_sqlite3.dll"
 )
 
 foreach ($dep in $Dependencies) {